Refactored extenders package
