+++ internal/gardener/shoot/extenders/annotations.go
@@ -1,10 +1,10 @@
-package extender
+package extenders
 
 import (
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
 )
 
-func AnnotationsExtender(imv1.RuntimeShoot, *gardener.Shoot) error {
+func ExtendWithAnnotations(imv1.RuntimeShoot, *gardener.Shoot) error {
 	return nil
 }
