+++ internal/gardener/shoot/extenders/extender.go
@@ -1,8 +1,8 @@
-package extender
+package extenders
 
 import (
 	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
 )
 
-type Extender func(imv1.RuntimeShoot, *gardener.Shoot) error
+type Extend func(imv1.RuntimeShoot, *gardener.Shoot) error
